requirements.txt
discord.py-self
requests